requirements.txt
# ============================================================
# TokenShield Requirements — Updated with 2FA & TOTP Support
# ============================================================

# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Security & Authentication
PyJWT==2.8.0
bcrypt==4.1.2
Werkzeug==3.0.1

# Machine Learning
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# Database
SQLAlchemy==2.0.23

# === PHASE 2: 2FA & TOTP (Google Authenticator) ===
pyotp==2.9.0                 # TOTP generation & verification
qrcode[pil]==7.4.2           # QR code image generation for authenticator setup

# === PHASE 2: Email Verification (Flask-Mail) ===
Flask-Mail==0.10.0           # Email sending for security alerts

# === PHASE 2: Google OAuth Sign-In ===
google-auth==2.28.0          # Google OAuth token verification
google-auth-oauthlib==1.2.0  # OAuth flow helpers
requests-oauthlib==1.3.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development & Testing
pytest==7.4.4
pytest-flask==1.3.0
